pyproject.toml
[project]
name = "dictator"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.4.0",
    "inflect>=7.5.0",
    "loguru>=0.7.3",
    "mysql-connector-python>=9.1.0",
    "pandas>=2.2.3",
    "pydantic-settings>=2.13.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentry-sdk>=2.20.0",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "pytest>=8.4.1",
    "ruff>=0.11.2",
    "ty>=0.0.14",
]
